refactor: extract Meteoroid class to its own module
- Moved the complex Meteoroid class to sprites/meteoroid.py
- Added import in sprites/__init__.py
- Verified all meteoroid behavior: spawning, movement, rotation, and splitting
