+++ sprites/__init__.py
@@ -1,2 +1,3 @@
 from .sprites import *
 from .starfield import Starfield
+from .meteoroid import Meteoroid
